move CreoFeatureCollection tests to dedicated file #25
